api/requirements.txt
azure-functions
openai
python-dotenv
requests
azure-identity

